requirements.txt
telebot
python-dotenv
pymongo
google-generativeai>=0.3.2
dnspython  # Resolve problemas de conexão do MongoDB